agentic-ai-business-case/agents/requirements.txt
strands-agents
strands-agents-tools
boto3>=1.34.90
botocore>=1.34.90
pyyaml>=6.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0
python-pptx>=0.6.23
python-docx>=0.8.11
reportlab>=4.0.0
pypdf>=4.0.0
